Moving sqlite related files to a separated folder to be better organized
