+++ src/dbsync/CMakeLists.txt
@@ -15,6 +15,7 @@ set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/lib)
 set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/bin)
 
 include_directories(${CMAKE_SOURCE_DIR}/include/)
+include_directories(${CMAKE_SOURCE_DIR}/src/)
 include_directories(${EXTERNAL_LIB}/sqlite/)
 include_directories(${EXTERNAL_LIB}/nlohmann/)
 include_directories(${EXTERNAL_LIB}/cJSON/)
@@ -22,7 +23,8 @@ link_directories(${EXTERNAL_LIB}/sqlite/)
 link_directories(${EXTERNAL_LIB}/cJSON/)
 
 file(GLOB DBSYNC_SRC
-    "${CMAKE_SOURCE_DIR}/src/*.cpp")
+    "${CMAKE_SOURCE_DIR}/src/*.cpp"
+    "${CMAKE_SOURCE_DIR}/src/sqlite/*.cpp")
 
 add_library(dbsync SHARED 
     "${DBSYNC_SRC}" )
