+++ src/dbsync/tests/CMakeLists.txt
@@ -5,6 +5,7 @@ project(unit_tests)
 include_directories(${EXTERNAL_LIB}/googletest/googletest/include/)
 include_directories(${EXTERNAL_LIB}/googletest/googlemock/include/)
 include_directories(${CMAKE_SOURCE_DIR}/src/)
+include_directories(${CMAKE_SOURCE_DIR}/src/sqlite/)
 link_directories(${EXTERNAL_LIB}/googletest/lib/)
