+++ src/dbsync/tests/sqlite/CMakeLists.txt
@@ -5,11 +5,12 @@ project(sqlite_unit_test)
 set(CMAKE_CXX_FLAGS_DEBUG "-g --coverage")
 
 include_directories(${CMAKE_SOURCE_DIR}/include/)
+include_directories(${CMAKE_SOURCE_DIR}/src/sqlite/)
 file(GLOB SQLITE_UNIT_TEST_SRC
     "*.cpp")
 
 file(GLOB SQLITE_SRC
-    "${CMAKE_SOURCE_DIR}/src/sqlite_wrapper.cpp")
+    "${CMAKE_SOURCE_DIR}/src/sqlite/sqlite_wrapper.cpp")
 
 add_executable(sqlite_unit_test 
     "${SQLITE_UNIT_TEST_SRC}"
